refactor(settings): 설정 모달 UI 섹션 분리
